Avoid overflow by converting to long

+++ server/src/main/java/org/apache/iotdb/db/utils/datastructure/TVList.java
@@ -82,11 +82,9 @@ public abstract class TVList {
   public static long tvListArrayMemSize(TSDataType type) {
     long size = 0;
     // time size
-    size +=
-        PrimitiveArrayManager.ARRAY_SIZE * 8;
+    size += (long) PrimitiveArrayManager.ARRAY_SIZE * 8L;
     // value size
-    size +=
-        PrimitiveArrayManager.ARRAY_SIZE * type.getDataTypeSize();
+    size += (long) PrimitiveArrayManager.ARRAY_SIZE * (long) type.getDataTypeSize();
     return size;
   }
